% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/subset_pots_image.r
\name{subset_pots}
\alias{subset_pots}
\title{Subset pots v_0.2}
\usage{
subset_pots(file_image = NA, dir = "./", outdir = "./",
  file_txt = paste(unlist(strsplit(paste(file_image), "[.]"))[1], "txt", sep =
  "."), barcode_position = "outside", pots_positions = c(1:16),
  size = 325)
}
\description{
This functions takes a single image or all the images contained in a directory and subset all the pots contained in the image
By default this function only takes JPG images (case insensitive).
It is mandatory that the last character of  dir and outputdir is "/" as in "/path1/path2/path3/"
The default dir and outputdir is "./"
User should specify the position of the barcode modyfing the parameter -barcode_position- Default outside
-pots_positions- defined which specific pots of the image the user want to subset. Default 1-16
-size- defined the length of the square surrounding the pot to subser. Default 350
Examples of Usage
}
\details{
Single image in ./ : subset_pots("my.image.jpg")
Single image in other path :subset_pots("/path/file.jpg")
Full directory :subset_pots(dir="path/")
Full directory ./ : subset_pots()
}

